Adding my open-source Java API
I have also written a Simple Blockchain API (SBA) in Java. It is very easy to use and great for beginners. There is also an example which shows the user how to use my API to create a blockchain. I would be very grateful if you could add this API to your Build-Your-Own list.
BowStar