2D game development in Java
I found this gem on Youtube after searching again and again, this video list helped me a lot to understand basic 2D game features and of course after this video serious it helped me to try my own 2D game. so I think it's better to share this with the community as me like guys would be take advantage of this.