Add Table of Contents heading + Back to Top link
The README is very large and includes many categories (AI, Blockchain, Database, etc.). 
Added a clear "Table of Contents" heading near the top so readers immediately understand that the list is navigational.

Also added a "Back to Top" link at the end of the README to make scrolling easier.

This improves usability for users browsing the long README and addresses issue #1599.